python_app/pyproject.toml
[project]
name = "algo-trading-py"
version = "0.1.0"
description = "Python migration baseline for algo-trading"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.11.0",
  "typer>=0.16.0",
  "uvicorn>=0.35.0",
]

[project.scripts]
algo-trading-py = "algo_trading_py.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
